Implement Daily Training System with Dice Mechanic (#25)
* feat: update training stats page with new roleplay submission mode and theme color adjustments

* feat: add action constants and local storage keys; refactor training-related functions

* feat: enhance Daily Training Config with dice roller and target confirmation logic; improve styling and scrollbar handling

* feat(training): implement multi-target training with early fail/success detection

- Updated `TrainWithAdmin` component to handle multiple training targets (5) instead of a single target.
- Introduced early failure and early win modals based on roll results.
- Modified training progress saving to allow partial progress after each roll.
- Added new functions in `dailyTrainingDice` service for managing multiple targets, including saving drafts and confirming targets.
- Updated Firestore structure to accommodate multiple targets and their confirmation status.
- Enhanced error handling and user feedback throughout the training process.
- Added `adminOnly` property to `CampLocation` type for future admin-specific features.

* feat: Add Training Roleplay Submission page and integrate ticket system for character requirements

- Implemented TrainingRoleplaySubmission component for submitting roleplay tasks.
- Added ticket functionality to reduce character requirements for submissions.
- Enhanced user experience with loading states and error handling.
- Updated UserDailyProgress interface to include tickets.
- Modified submitTrainingRoleplay service to accept tickets as a parameter.

* feat: add TrainingPracticeModal component for managing practice rooms

refactor: update PvP page to utilize Arena component

style: center align mode elements in Stats.scss

fix: enable PvP mode selection in Stats component

style: adjust error display in TrainWithAdmin

refactor: optimize data loading in TrainWithAdmin component

feat: enhance TrainingRoleplaySubmission with PvP practice handling

feat: implement savePracticeProgress function for tracking PvP practice

fix: update BattleRoom type to include practiceMode field

* Enhance Training Practice Modal and PvP Functionality

- Updated TrainingPracticeModal styles for improved hover effects and loading indicators.
- Refactored TrainingPracticeModal component to handle loading states more effectively.
- Introduced loading blocks and text to indicate ongoing processes in TrainingPracticeModal.
- Enhanced PvP page to manage user training states and display appropriate messages based on training status.
- Added logic to prevent starting new training sessions if the user has pending tasks or ongoing PvP sessions.
- Improved error handling and user feedback in TrainWithAdmin and TrainingRoleplaySubmission components.
- Updated daily training services to support additional fields for Google Sheets integration.
- Ensured proper state management for live practice sessions and training tasks across components.

* feat(training): enhance refund and upgrade functionalities with modals and animations

* feat(training): implement refund all stats functionality and UI enhancements

* feat(training): refactor date formatting functions for consistency

* feat: enhance lobby and training grounds functionality

- Added validation to ensure rooms are only joined if their status is WAITING in Lobby.
- Improved TrainingGrounds component to handle existing PvP rooms and modal states more effectively.
- Updated TrainingPracticeModal to navigate to the PvP page after finalizing a practice room.
- Refactored PvP component to handle direct arena access and loading states.
- Enhanced TrainWithAdmin component to manage training tasks and PvP room states.
- Added new utility function to check success based on individual targets in daily training.
- Updated Firestore and Realtime Database interactions for training quotas and progress tracking.
- Improved styling for training components to reflect task statuses.

* feat(training): add arena and practice constants, refactor related components

* Refactor training and harvest services

- Updated EarlyFailModal styles to enforce button background color.
- Modified TrainingRoleplaySubmission to use new TrainingTask interface and updated logic for handling training tasks.
- Removed unnecessary console.error statements in firebaseAnonymous, fetchHarvest, trainingPoints, and upgradeStats services.
- Enhanced dailyTrainingDice service with new TrainingTask structure and streamlined training result submission.
- Introduced canUserTrain function for comprehensive training validation.
- Created new trainingTask types for better organization and re-exported harvest-related types.

* feat: Add new Tweets icon component and update existing icons for consistency

fix: Remove unused Firebase database imports in Arena and TrainingGrounds

style: Update Skull and Sword icons for better SVG structure and styling

refactor: Change date format in HarvestRecordCard to short month

style: Enhance SubmissionCard styles for better layout and focus state

fix: Update TrainingGrounds logic to handle user training tasks more effectively

feat: Implement recheck functionality in TrainingRoleplaySubmission with new Alert icon

fix: Trim tweet text in harvest approval service for cleaner output

refactor: Adjust date formatting in verifyTrainingTask for consistency

* feat(training): enhance training point awarding logic and update UI components

+++ src/pages/StrawberryFields/components/HarvestRecordCard/HarvestRecordCard.tsx
@@ -25,7 +25,7 @@ export default function HarvestRecordCard({ submission, characterMap }: HarvestR
 
   const date = new Date(submission.submittedAt).toLocaleDateString(
     lang === LANGUAGE.ENGLISH ? 'en-US' : 'th-TH',
-    { year: 'numeric', month: 'long', day: 'numeric' }
+    { year: 'numeric', month: 'short', day: 'numeric' }
   );
 
   const submitter = characterMap[submission.characterId.toLowerCase()] || null;
